Fix getOrdinal() prototype
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@2755 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src/x10/lang/floatArray.java
@@ -36,7 +36,7 @@ abstract public class floatArray extends x10Array {
 	 * region.
 	 */
 	abstract public float get(point/*(region)*/ p);
-	abstract public float getOrdinal(float p);
+	abstract public float getOrdinal(int p);
 	abstract /*value*/ public float get(int p);
 	abstract /*value*/ public float get(int p, int q);
 	abstract /*value*/ public float get(int p, int q, int r);
